added seperate file for ProgressBar
